Move DataFrameRepartitionSuite to DataFrameSuite
